+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 #### Breaking Changes
 * `NOTAM::Schedule.parse` now returns an array of `NOTAM_Schedule` instances
-  instead of just a single one (to model date and time ranges across midnight).
+  instead of just a single one.
 
 #### Additions
+* Support for datetime ranges (i.e. `1 APR 2000-20 MAY 2000`) as well as times
+  across midnight (i.e. `1 APR 1900-0500`) on D items.
 * Wrap all exceptions raised while parsing items.
 
 ## 0.1.3